Fixed get_all_sites issue.
Craigslist must have added an extra '/' to the href for their official sites ex: <li><a href="//longisland.craigslist.org/">long island</a></li>. 
Noticed this when attempting to use the library (it's awesome). A not a real site error was thrown when I was trying to use the library, so I started to investigate the get_all_sites() function in sites.py. Added an extra '/' to the split and voila it worked.