Link filter option with its numerical value
Previously, filters from utils.get_list_filters returned a list of filters in <div class="search-attribute">. Numerical values that are tied to these filters (and ultimately concatenated to the query url) were assumed to increment by 1 from 1. This is not always the case, which in turn leads to an invalid filter code concatenated to the query url.

These numerical codes are actually tied to their filters; therefore, this commit links the filter option (e.g. 'silver') with its numerical value (e.g. 8).